fixed bug that prevented the same TileSpec instance from rendering correctly twice.  Renamed TileSpec.createTransformList() to getTransformList() because the method does not necessarily create a new instance.  Exactly this was the reason of the bug:
If the TileSpec had a list of Transforms, createTransfomrList() would
return this instance instead of a copy.  The instance is currently held
as a transient private member for later reuse.  Render.render(...) would
then add bounding box, pixel offset, mipmap scaling and result scaling
related transforms to this list, modifying it in place.
I have modified render(...) to copy the list, renamed the method, and
added care-demanding documentation.

Explanation: